WIP: update few missing deps and move tools around
