+++ packages/cli-doctor/package.json
@@ -8,10 +8,18 @@
   },
   "types": "build/index.d.ts",
   "dependencies": {
-    "@react-native-community/cli-tools": "^6.0.0-rc.0",
+    "@react-native-community/cli-tools": "^6.2.0",
     "chalk": "^3.0.0",
+    "envinfo": "^7.7.2",
     "hermes-profile-transformer": "^0.0.6",
-    "ip": "^1.1.5"
+    "ip": "^1.1.5",
+    "command-exists": "^1.2.8",
+    "prompts": "^2.4.0",
+    "semver": "^6.3.0",
+    "strip-ansi": "^5.2.0",
+    "ora": "^6.0.0",
+    "execa": "^1.0.0",
+    "wcwidth": "^1.0.1"
   },
   "files": [
     "build",
@@ -19,8 +27,12 @@
     "!*.map"
   ],
   "devDependencies": {
+    "@types/command-exists": "^1.2.0",
     "@react-native-community/cli-types": "^6.0.0",
-    "@types/ip": "^1.1.0"
+    "@types/ip": "^1.1.0",
+    "@types/semver": "^6.0.2",
+    "@types/wcwidth": "^1.0.0",
+    "@types/prompts": "^2.0.9"
   },
   "homepage": "https://github.com/react-native-community/cli/tree/master/packages/cli-doctor",
   "repository": {
