+++ packages/cli-doctor/src/commands/__tests__/info.test.ts
