+++ packages/cli-doctor/src/commands/doctor.ts
@@ -1,0 +1,21 @@
+import doctor from './doctor';
+
+export default {
+    func: doctor,
+    detached: true,
+    name: 'doctor',
+    description:
+      '[EXPERIMENTAL] Diagnose and fix common Node.js, iOS, Android & React Native issues.',
+    options: [
+      {
+        name: '--fix',
+        description: 'Attempt to fix all diagnosed issues.',
+      },
+      {
+        name: '--contributor',
+        description:
+          'Add healthchecks required to installations required for contributing to React Native.',
+      },
+    ],
+  };
+
