+++ packages/cli-doctor/src/tools/checkInstallation.ts
