+++ packages/cli-doctor/src/tools/envinfo.ts
