+++ packages/cli-doctor/src/tools/healthchecks/__tests__/androidHomeEnvVariable.test.ts
