+++ packages/cli-doctor/src/tools/healthchecks/__tests__/androidNDK.test.ts
@@ -1,5 +1,5 @@
 import androidNDK from '../androidNDK';
-import getEnvironmentInfo from '@react-native-community/cli/src/tools/envinfo';
+import getEnvironmentInfo from '../../envinfo';
 import {EnvironmentInfo} from '@react-native-community/cli-types';
 import {NoopLoader} from '@react-native-community/cli/src/tools/loader';
