+++ packages/cli-doctor/src/tools/healthchecks/androidHomeEnvVariable.ts
