+++ packages/cli-doctor/src/tools/healthchecks/androidNDK.ts
