+++ packages/cli-doctor/src/tools/healthchecks/androidSDK.ts
