+++ packages/cli-doctor/src/tools/healthchecks/androidStudio.ts
