+++ packages/cli-doctor/src/tools/healthchecks/cocoaPods.ts
