+++ packages/cli-doctor/src/tools/healthchecks/common.ts
