+++ packages/cli-doctor/src/tools/healthchecks/index.ts
