+++ packages/cli-doctor/src/tools/healthchecks/jdk.ts
