+++ packages/cli-doctor/src/tools/healthchecks/packageManagers.ts
