+++ packages/cli-doctor/src/tools/healthchecks/watchman.ts
