+++ packages/cli-doctor/src/tools/healthchecks/xcode.ts
