+++ packages/cli-doctor/src/tools/printFixOptions.ts
