+++ packages/cli-doctor/src/tools/runAutomaticFix.ts
