+++ packages/cli-doctor/src/tools/types.ts
