+++ packages/cli-doctor/src/tools/versionRanges.ts
