+++ packages/cli/package.json
@@ -25,19 +25,17 @@
   },
   "dependencies": {
     "@react-native-community/cli-debugger-ui": "^6.0.0-rc.0",
+    "@react-native-community/cli-doctor": "^6.0.0",
     "@react-native-community/cli-hermes": "^6.3.0",
     "@react-native-community/cli-plugin-metro": "^6.4.0",
     "@react-native-community/cli-server-api": "^6.4.0",
     "@react-native-community/cli-tools": "^6.2.0",
     "@react-native-community/cli-types": "^6.0.0",
-    "@react-native-community/cli-doctor": "^6.0.0",
     "appdirsjs": "^1.2.4",
     "chalk": "^4.1.2",
-    "command-exists": "^1.2.8",
     "commander": "^2.19.0",
     "cosmiconfig": "^5.1.0",
     "deepmerge": "^3.2.0",
-    "envinfo": "^7.7.2",
     "execa": "^1.0.0",
     "find-up": "^4.1.0",
     "fs-extra": "^8.1.0",
@@ -53,15 +51,12 @@
     "prompts": "^2.4.0",
     "semver": "^6.3.0",
     "serve-static": "^1.13.1",
-    "strip-ansi": "^5.2.0",
-    "sudo-prompt": "^9.0.0",
-    "wcwidth": "^1.0.1"
+    "sudo-prompt": "^9.0.0"
   },
   "peerDependencies": {
     "react-native": "*"
   },
   "devDependencies": {
-    "@types/command-exists": "^1.2.0",
     "@types/cosmiconfig": "^5.0.3",
     "@types/fs-extra": "^8.1.0",
     "@types/glob": "^7.1.1",
@@ -71,7 +66,6 @@
     "@types/minimist": "^1.2.0",
     "@types/prompts": "^2.0.9",
     "@types/semver": "^6.0.2",
-    "@types/wcwidth": "^1.0.0",
     "slash": "^3.0.0",
     "snapshot-diff": "^0.7.0"
   },
